exporting LDAP access beans to their own classes
